backend/requirements.txt
# Django==4.2.9
# djangorestframework==3.14.0
# djangorestframework-simplejwt==5.3.1
# django-cors-headers==4.3.1
# psycopg2-binary==2.9.9
# python-decouple==3.8
# Pillow==10.4.0
# twilio==8.10.0
# django-filter==23.5
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
twilio==8.10.0
django-filter==24.3